moved Example.js into example folder
