+++ src/pages/index.js
@@ -3,7 +3,7 @@ import { Link } from "gatsby"
 import Layout from "../components/layout"
 import Image from "../components/image"
 import SEO from "../components/seo"
-import ExampleComponent from '../components/Example';
+import ExampleComponent from '../components/example/Example';
 
 const IndexPage = () => (
   <Layout>
